Move Commenting code into trait
